backend/requirements.txt
Flask
flask-cors
transformers
torch
sentencepiece
PyMuPDF
pytesseract
pdf2image
Pillow
python-docx
pytube
uuid
